create summaryRangeMapper. implement getAccountsummaryRange. refactor income amounts to be bigDecimal instead of double.

+++ aggregation-model/src/main/java/com/quicken/aggregation_model/dto/SummaryRangeDto.java
@@ -1,5 +1,6 @@
 package com.quicken.aggregation_model.dto;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 import lombok.AllArgsConstructor;
@@ -8,9 +9,9 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class SummaryRangeDto {
-    double income;
-    double expenses;
-    double net;
+    BigDecimal income;
+    BigDecimal expenses;
+    BigDecimal net;
     Date startDate;
     Date endDate;
 }
